Internal interface

Documentation for RandomFields.jl's internal interface.

Contents

Contents

Functions

Types

Index

Functions

RandomFields.apply! — Method

apply!(result, operand, generator)

Apply the symmetrized modified Helmholtz operator underlying the Gaussian random field generator to an operand field and write in-place to result field.

RandomFields.bounding_pair — Function

bounding_pair(m, i, j, k)

The two indices, in (near, far) order, of the flipped-location evaluation points that bound a field's own location at indices (i, j, k) in dimension m. Assumes loc is the output (field's own) location in dimension m.

RandomFields.coloring_period — Method

coloring_period(N, R, is_periodic)

Smallest P >= 2R+1 dividing N, for a periodic axis and 2R+1 otherwise.

RandomFields.directional_operators — Method

directional_operators(grid, loc)

Construct directional differential operators for a field on a grid grid and with staggered grid locations tuple loc.

RandomFields.generate_white_noise! — Method

generate_white_noise!(
    field,
    noise,
    generator;
    mask_immersed
)

Generate white noise and write to field using standard normal variate in noise with size(noise) == size(field) and scaling by scale factor and reciprocal of square root of cell volumes cached in generator. If field is defined on a ImmersedBoundaryGrid and mask_immersed is true inactive immersed cells in the output will be masked to zero.

RandomFields.get_immersed_indices — Method

get_immersed_indices(field)

Get a vector of linear indices in to array corresponding to interior of field for cells which are inactive as they are fully immersed.

RandomFields.get_sparse_operator — Method

get_sparse_operator(
    result,
    operand,
    apply!,
    args;
    stencil_radius,
    verify,
    atol
)

Determine the sparse matrix representation (A, b) of the linear (affine) operator implicitly defined by apply!.

The sparse matrix A and affine offset vector b are computed such that result_vec .= A * vec(interior(operand)) + b corresponds to apply!(result, operand, args...) with result_vec == vec(interior(result)). A is restricted to acting on interior grid points and with ordering corresponding to the contiguous layout of result and operand fields data arrays. For a purely linear operator (homogeneous boundary conditions), b is all zeros.

apply! must be linear and spatially local with support within stencil_radius (in L∞ / Chebyshev distance). If verify=true a second probe is performed with stencil_radius + 1 and if the resulting computed A differs an error is raised.

Requires (2*stencil_radius+1)^3 calls to apply!, independent of grid size, or 2 * (2*stencil_radius+2)^3 + 12 * (stencil_radius+1) calls if verify=true.

RandomFields.masked_evaluate — Method

masked_evaluate(op, m, i, j, k, grid, operand)

Zero out a directional derivative op evaluated at (i,j,k) on grid with operand if either of the two cells bounding that point (along dimension m) is solid.

RandomFields.regularize_operator — Method

regularize_operator(A; ε)

Regularize a symmetric stencil operator A for which some rows / columns are all zero due to, for example, presence of inactive immersed cells in the corresponding fields, by adding an arbitrary value ε to the corresponding diagonal entries, returning the reguarlized matrix A_reg and the vector of inactive indices inactive.

RandomFields.second_derivative_term — Function

second_derivative_term(i, j, k, grid, op, far, near, flux_far, flux_near)

Compute second derivative term for a field on grid grid with direction differential operator op at indices (i, j, k) given the computed flux terms flux_near and flux_far at near and far offset index tuples far and near.

RandomFields.shift_index — Function

shift_index(m, i, j, k, offset)

Shift indices (i, j, k) by offset cells along dimension m.

RandomFields.solve! — Function

solve!(solution, rhs, solver)

Apply the inverse of the linear operator solved for by solver to the right-hand side rhs and write the solution in-place to solution, where rhs and solution are both fields.

Solves A * x = b where A is the linear system solved by solver, b is the vectorization of the rhs field and x is the vectorization of the solution field.

RandomFields.solve_adjoint! — Function

solve_adjoint!(solution, rhs, solver)

Apply the inverse of the adjoint of the linear operator solved for by solver to the right-hand side rhs and write the solution in-place to solution, where rhs and solution are both fields.

Solves A' * x = b where A' is the adjoint of the linear system solved by solver, b is the vectorization of the rhs field and x is the vectorization of the solution field.

RandomFields.solve_sqrt! — Function

solve_sqrt!(solution, rhs, solver)

Apply the inverse of the square-root of linear operator solved for by solver to the right-hand side rhs and write the solution in-place to solution, where rhs and solution are both fields.

Solves sqrt(A) * x = b where A is the linear system solved by solver and sqrt(A) is a matrix such that sqrt(A) * sqrt(A)' === A, b is the vectorization of the rhs field and x is the vectorization of the solution field.

RandomFields.solve_sqrt_adjoint! — Function

solve_sqrt_adjoint!(solution, rhs, solver)

Apply the inverse of the adjoint of the square-root of linear operator solved for by solver to the right-hand side rhs and write the solution in-place to solution, where rhs and solution are both fields.

Solves sqrt(A)' * x = b where A is the linear system solved by solver and sqrt(A) is a matrix such that sqrt(A) * sqrt(A)' === A, b is the vectorization of the rhs field and x is the vectorization of the solution field.

RandomFields.weighted_second_derivative_sum — Method

weighted_second_derivative_sum(
    i,
    j,
    k,
    grid,
    operand,
    weights,
    operators
)

Compute the weighted second derivative sum for a field operand on grid grid at indices (i, j, k) with per-dimension derivative weights weights and directional derivative operators operators.

RandomFields.zero! — Method

zero!(field)

Fill field with zeros in-place.

Types

RandomFields.NonSeparableDirectionalOperator — Type

struct NonSeparableDirectionalOperator{M, I, A, V} <: RandomFields.AbstractDirectionalOperator{M, I}

Directional differential operator on a dimension M for grids with non-separable metrics.

Fields

inner::Any: Derivative operation at dimension M flipped operand location

area::Any: Face area operation at dimension M flipped operand location

volume::Any: Cell volume operation at operand location

RandomFields.SeparableDirectionalOperator — Type

struct SeparableDirectionalOperator{M, I, D} <: RandomFields.AbstractDirectionalOperator{M, I}

Directional differential operator on a dimension M for grids with separable metrics.

Fields

inner::Any: Derivative operation at dimension M flipped operand location

delta::Any: Grid spacing operation at operand location

Index

RandomFields.NonSeparableDirectionalOperator

RandomFields.SeparableDirectionalOperator

RandomFields.apply!

RandomFields.bounding_pair

RandomFields.coloring_period

RandomFields.directional_operators

RandomFields.generate_white_noise!

RandomFields.get_immersed_indices

RandomFields.get_sparse_operator

RandomFields.masked_evaluate

RandomFields.regularize_operator

RandomFields.second_derivative_term

RandomFields.shift_index

RandomFields.solve!

RandomFields.solve_adjoint!

RandomFields.solve_sqrt!

RandomFields.solve_sqrt_adjoint!

RandomFields.weighted_second_derivative_sum

RandomFields.zero!